refactor(scripts): use dependency metadata parsers
