+++ scripts/gh-read.ts
@@ -1,11 +1,17 @@
 // Gh Read script supports OpenClaw repository automation.
-import { execFileSync, spawnSync } from "node:child_process";
+import { spawnSync } from "node:child_process";
 import { createPrivateKey, createSign } from "node:crypto";
 import { readFileSync } from "node:fs";
 import { pathToFileURL } from "node:url";
 import { expectDefined } from "../packages/normalization-core/src/expect.js";
 import { readBoundedResponseText } from "./lib/bounded-response.ts";
 import { parseStrictIntegerOption } from "./lib/dev-tooling-safety.ts";
+import {
+  normalizeGitHubRepo as normalizeRepo,
+  resolveGitHubRepoFromOrigin,
+} from "./lib/github-repo.ts";
+
+export { normalizeRepo };
 
 const APP_ID_ENV = "OPENCLAW_GH_READ_APP_ID";
 const KEY_FILE_ENV = "OPENCLAW_GH_READ_PRIVATE_KEY_FILE";
@@ -63,23 +69,6 @@ export function parseRepoArg(args: string[]): string | null {
     }
   }
   return null;
-}
-
-export function normalizeRepo(value: string | null | undefined): string | null {
-  const trimmed = value?.trim();
-  if (!trimmed) {
-    return null;
-  }
-
-  const withoutProtocol = trimmed.replace(/^[a-z]+:\/\//i, "");
-  const withoutHost = withoutProtocol.replace(/^(?:[^@/]+@)?github\.com[:/]/i, "");
-  const normalized = withoutHost.replace(/\.git$/i, "").replace(/^\/+|\/+$/g, "");
-  const parts = normalized.split("/").filter(Boolean);
-  if (parts.length < 2) {
-    return null;
-  }
-
-  return `${parts[parts.length - 2]}/${parts[parts.length - 1]}`;
 }
 
 export function parsePermissionKeys(raw: string | null | undefined): string[] {
@@ -147,11 +136,7 @@ function resolveRepo(args: string[]): string | null {
   }
 
   try {
-    const remote = execFileSync("git", ["config", "--get", "remote.origin.url"], {
-      encoding: "utf8",
-      stdio: ["ignore", "pipe", "ignore"],
-    }).trim();
-    return normalizeRepo(remote);
+    return resolveGitHubRepoFromOrigin();
   } catch {
     return null;
   }
