+++ scripts/label-open-issues.ts
@@ -7,6 +7,7 @@ import { pathToFileURL } from "node:url";
 import { isRecord } from "../src/utils.js";
 import { readBoundedResponseText as readBoundedBodyText } from "./lib/bounded-response.ts";
 import { parseStrictIntegerOption } from "./lib/dev-tooling-safety.ts";
+import { resolveGitHubRepoFromOrigin } from "./lib/github-repo.ts";
 
 function writeStdoutLine(message = ""): void {
   process.stdout.write(`${message}\n`);
@@ -466,33 +467,8 @@ function runGh(args: string[]): string {
 }
 
 function resolveRepo(): RepoInfo {
-  const remote = execFileSync("git", ["config", "--get", "remote.origin.url"], {
-    encoding: "utf8",
-  }).trim();
-
-  if (!remote) {
-    throw new Error("Unable to determine repository from git remote.");
-  }
-
-  const normalized = remote.replace(/\.git$/, "");
-
-  if (normalized.startsWith("git@github.com:")) {
-    const slug = normalized.replace("git@github.com:", "");
-    const [owner, name] = slug.split("/");
-    if (owner && name) {
-      return { owner, name };
-    }
-  }
-
-  if (normalized.startsWith("https://github.com/")) {
-    const slug = normalized.replace("https://github.com/", "");
-    const [owner, name] = slug.split("/");
-    if (owner && name) {
-      return { owner, name };
-    }
-  }
-
-  throw new Error(`Unsupported GitHub remote: ${remote}`);
+  const [owner, name] = resolveGitHubRepoFromOrigin().split("/") as [string, string];
+  return { owner, name };
 }
 
 function fetchIssuePage(repo: RepoInfo, after: string | null): IssuePage {
