+++ scripts/sync-labels.ts
@@ -2,6 +2,8 @@
 import { execFileSync } from "node:child_process";
 import { readFileSync } from "node:fs";
 import { resolve } from "node:path";
+import { parse } from "yaml";
+import { resolveGitHubRepoFromOrigin } from "./lib/github-repo.ts";
 
 type RepoLabel = {
   name: string;
@@ -48,15 +50,10 @@ const EXTRA_LABELS = [
   "size: XL",
   "beta-blocker",
 ] as const;
-const labelNames = [
-  ...new Set([...extractLabelNames(readFileSync(configPath, "utf8")), ...EXTRA_LABELS]),
-];
+const labelerConfig = parse(readFileSync(configPath, "utf8")) as Record<string, unknown>;
+const labelNames = [...new Set([...Object.keys(labelerConfig), ...EXTRA_LABELS])];
 
-if (!labelNames.length) {
-  throw new Error("labeler.yml must declare at least one label.");
-}
-
-const repo = resolveRepo();
+const repo = resolveGitHubRepoFromOrigin();
 const existing = fetchExistingLabels(repo);
 
 const missing = labelNames.filter((label) => !existing.has(label));
@@ -84,26 +81,6 @@ for (const label of missing) {
   console.log(`Created label: ${label}`);
 }
 
-function extractLabelNames(contents: string): string[] {
-  const labels: string[] = [];
-  for (const line of contents.split("\n")) {
-    if (!line.trim() || line.trimStart().startsWith("#")) {
-      continue;
-    }
-    if (/^\s/.test(line)) {
-      continue;
-    }
-    const match = line.match(/^(["'])(.+)\1\s*:/) ?? line.match(/^([^:]+):/);
-    if (match) {
-      const name = (match[2] ?? match[1] ?? "").trim();
-      if (name) {
-        labels.push(name);
-      }
-    }
-  }
-  return labels;
-}
-
 function resolveLabelMetadata(label: string): { color: string; description?: string } {
   const extraMetadata = EXTRA_LABEL_METADATA.get(label);
   if (extraMetadata) {
@@ -111,26 +88,6 @@ function resolveLabelMetadata(label: string): { color: string; description?: str
   }
   const prefix = label.includes(":") ? label.slice(0, label.indexOf(":")).trim() : label.trim();
   return { color: COLOR_BY_PREFIX.get(prefix) ?? "ededed" };
-}
-
-function resolveRepo(): string {
-  const remote = execFileSync("git", ["config", "--get", "remote.origin.url"], {
-    encoding: "utf8",
-  }).trim();
-
-  if (!remote) {
-    throw new Error("Unable to determine repository from git remote.");
-  }
-
-  if (remote.startsWith("git@github.com:")) {
-    return remote.replace("git@github.com:", "").replace(/\.git$/, "");
-  }
-
-  if (remote.startsWith("https://github.com/")) {
-    return remote.replace("https://github.com/", "").replace(/\.git$/, "");
-  }
-
-  throw new Error(`Unsupported GitHub remote: ${remote}`);
 }
 
 function fetchExistingLabels(repoLocal: string): Map<string, RepoLabel> {
